requirements.txt
future==0.16.0
python-whois==0.6.5
requests==2.13.0
